refactor(test): split preview tests by preview family
